+++ library/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2022-11-19
+
+### Added
+- added a file provider at @gorgon/file-provider
+### Fixed
+- fixed a bug where when you cached something permanently (false) with the non default cache it would use the default provider
+
 ## [1.3.1] - 2022-11-14
 
 ### Added